# Conflicts:
#	.skills/SKILL.md
#	docs/sd-card-fonts.md
#	lib/Epub/Epub/BookMetadataCache.h
#	lib/I18n/translations/belarusian.yaml
#	lib/I18n/translations/catalan.yaml
#	lib/I18n/translations/czech.yaml
#	lib/I18n/translations/danish.yaml
#	lib/I18n/translations/dutch.yaml
#	lib/I18n/translations/english.yaml
#	lib/I18n/translations/finnish.yaml
#	lib/I18n/translations/french.yaml
#	lib/I18n/translations/german.yaml
#	lib/I18n/translations/hungarian.yaml
#	lib/I18n/translations/italian.yaml
#	lib/I18n/translations/kazakh.yaml
#	lib/I18n/translations/lithuanian.yaml
#	lib/I18n/translations/polish.yaml
#	lib/I18n/translations/portuguese.yaml
#	lib/I18n/translations/romanian.yaml
#	lib/I18n/translations/russian.yaml
#	lib/I18n/translations/slovenian.yaml
#	lib/I18n/translations/spanish.yaml
#	lib/I18n/translations/swedish.yaml
#	lib/I18n/translations/turkish.yaml
#	lib/I18n/translations/ukrainian.yaml
#	lib/JpegToBmpConverter/JpegToBmpConverter.cpp
#	lib/KOReaderSync/KOReaderCredentialStore.cpp
#	lib/KOReaderSync/KOReaderDocumentId.cpp
#	lib/PngToBmpConverter/PngToBmpConverter.cpp
#	lib/Serialization/Serialization.h
#	lib/hal/HalStorage.h
#	src/CrossPointSettings.cpp
#	src/CrossPointState.cpp
#	src/FontInstaller.cpp
#	src/RecentBooksStore.cpp
#	src/WifiCredentialStore.cpp
#	src/activities/settings/FontDownloadActivity.cpp
#	src/network/CrossPointWebServer.cpp
#	src/network/CrossPointWebServer.h
#	src/network/HttpDownloader.cpp
#	src/network/WebDAVHandler.cpp
#	src/network/WebDAVHandler.h
#	src/network/html/FilesPage.html

+++ src/activities/network/WifiSelectionActivity.h
@@ -80,8 +80,8 @@ class WifiSelectionActivity final : public Activity {
   int forgetPromptSelection = 0;
 
   // Connection timeout
-  static constexpr unsigned long CONNECTION_TIMEOUT_MS = 15000;
-  unsigned long connectionStartTime = 0;
+  static constexpr uint32_t CONNECTION_TIMEOUT_MS = 15000;
+  uint32_t connectionStartTime = 0;
 
   void renderNetworkList(const Rect* screen, const ThemeMetrics* metrics) const;
   void renderPasswordEntry(const Rect* screen, const ThemeMetrics* metrics) const;
